feat: add popup to home
